requirements.txt
Jinja2>=3.0
python-dotenv>=1.0
openai>=1.0
